Factor out panes within tabs
